Adjustments for Visual Studio 2013. Now using WiX instead of old setup project for installations. Still installed as an addin.

+++ dev/sqtpp/File.cpp
@@ -84,7 +84,7 @@ struct File::Data {
 	Data()
 	{
 		m_nRefCount         = 1;
-		m_nIncludeLevel     = -1;
+		m_nIncludeLevel     = (size_t)-1;
 		m_currentPosition   = 0;
 		m_pInternalStream   = NULL;
 		m_pExternalStream   = NULL;
